src/requirements.txt
discord.py>=2.5.2
yt-dlp>=2025.7.21
PyYAML>=6.0.2
PyNaCl>=1.5.0
requests>=2.32.4
python-dotenv>=1.1.1
aiohttp>=3.12.15
async-timeout>=4.0.3
aiofiles>=24.1.0